requirements.txt
tensorboard==1.14.0
tensorflow-gpu>=1.11.0,<=1.14.0
joblib==0.13.2
Keras==2.3.1
numpy==1.15.0
opencv-python==4.1.0.25
matplotlib==3.1.1
h5py==2.9.0
seaborn==0.9.0
pandas==0.25.3
Pillow==6.1.0
keras-retinanet==0.5.1
efficientnet==1.1.1
